split controller handler functions into files
