+++ src/controller/handlers/mod.rs
@@ -1,0 +1,6 @@
+pub mod audio;
+pub mod cacher;
+pub mod image_processor;
+pub mod lyrics;
+pub mod scanner;
+pub mod system_integration;
